pyproject.toml
[project]
name = "photocontest"
version = "0.1.0"
description = ""
authors = [
    {name = "Bogdan",email = "bgdnvlkvb7@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "django-stubs (>=5.2.2,<6.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "django-imagekit (>=5.0.0,<6.0.0)",
    "django-service-objects @ git+https://git.snpdev.ru/saltpepper/django-service-objects.git",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
